Update: Move Message to openai4s.types, add Message newtypes to Chat and Response, and reorganise tests
